+++ server/resolvers-custom.js
@@ -4,10 +4,12 @@ import fs from "fs";
 import { Kafka, Producer } from "kafkajs";
 import tracing, { Span } from "opentracing";
 import Arango from "./arango";
+import type { GraphQLRequestContext } from "./arango-collection";
 import type { QConfig } from "./config";
 import { ensureProtocol } from "./config";
 import path from 'path';
 import fetch from 'node-fetch';
+import QAuth from "./q-auth";
 import { QTracer } from "./tracer";
 
 function isObject(test: any): boolean {
@@ -33,11 +35,8 @@ type Request = {
     body: string,
 }
 
-type Context = {
+type GraphQLRequestContextEx = GraphQLRequestContext & {
     db: Arango,
-    config: QConfig,
-    remoteAddress?: string,
-    shared: Map<string, any>,
 }
 
 // Query
@@ -49,27 +48,30 @@ function info(): Info {
     };
 }
 
-async function getAccountsCount(_parent, args, context: Context): Promise<number> {
+async function getAccountsCount(_parent, args, context: GraphQLRequestContextEx): Promise<number> {
     const tracer = context.db.tracer;
     return QTracer.trace(tracer, 'getAccountsCount', async () => {
+        await context.auth.requireGrantedAccess(context.authToken || args.auth);
         const result: any = await context.db.query(`RETURN LENGTH(accounts)`, {});
         const counts = (result: number[]);
         return counts.length > 0 ? counts[0] : 0;
     }, QTracer.getParentSpan(tracer, context))
 }
 
-async function getTransactionsCount(_parent, args, context: Context): Promise<number> {
+async function getTransactionsCount(_parent, args, context: GraphQLRequestContextEx): Promise<number> {
     const tracer = context.db.tracer;
     return QTracer.trace(tracer, 'getTransactionsCount', async () => {
+        await context.auth.requireGrantedAccess(context.authToken || args.auth);
         const result: any = await context.db.query(`RETURN LENGTH(transactions)`, {});
         const counts = (result: number[]);
         return counts.length > 0 ? counts[0] : 0;
     }, QTracer.getParentSpan(tracer, context))
 }
 
-async function getAccountsTotalBalance(_parent, args, context: Context): Promise<String> {
+async function getAccountsTotalBalance(_parent, args, context: GraphQLRequestContextEx): Promise<String> {
     const tracer = context.db.tracer;
     return QTracer.trace(tracer, 'getAccountsTotalBalance', async () => {
+        await context.auth.requireGrantedAccess(context.authToken || args.auth);
         /*
         Because arango can not sum BigInt's we need to sum separately:
         hs = SUM of high bits (from 24-bit and higher)
@@ -93,7 +95,7 @@ async function getAccountsTotalBalance(_parent, args, context: Context): Promise
 
 // Mutation
 
-async function postRequestsUsingRest(requests: Request[], context: Context, span: Span): Promise<void> {
+async function postRequestsUsingRest(requests: Request[], context: GraphQLRequestContextEx, span: Span): Promise<void> {
     const config = context.config.requests;
     const url = `${ensureProtocol(config.server, 'http')}/topics/${config.topic}`;
     const response = await fetch(url, {
@@ -119,7 +121,7 @@ async function postRequestsUsingRest(requests: Request[], context: Context, span
     }
 }
 
-async function postRequestsUsingKafka(requests: Request[], context: Context, span: Span): Promise<void> {
+async function postRequestsUsingKafka(requests: Request[], context: GraphQLRequestContextEx, span: Span): Promise<void> {
     const ensureShared = async (name, createValue: () => Promise<any>) => {
         if (context.shared.has(name)) {
             return context.shared.get(name);
@@ -155,7 +157,7 @@ async function postRequestsUsingKafka(requests: Request[], context: Context, spa
     });
 }
 
-async function postRequests(_, args: { requests: Request[] }, context: Context): Promise<string[]> {
+async function postRequests(_, args: { requests: Request[], auth?: string }, context: GraphQLRequestContextEx): Promise<string[]> {
     const requests: ?(Request[]) = args.requests;
     if (!requests) {
         return [];
@@ -164,6 +166,7 @@ async function postRequests(_, args: { requests: Request[] }, context: Context):
     const tracer = context.db.tracer;
     return QTracer.trace(tracer, "postRequests", async (span: Span) => {
         span.setTag('params', requests);
+        await context.auth.requireGrantedAccess(context.authToken || args.auth);
         try {
             if (context.config.requests.mode === 'rest') {
                 await postRequestsUsingRest(requests, context, span);
@@ -176,7 +179,7 @@ async function postRequests(_, args: { requests: Request[] }, context: Context):
             throw error;
         }
         return requests.map(x => x.id);
-    }, QTracer.getParentSpan(tracer, context));
+    }, context.parentSpan);
 }
 
 const resolversCustom = {