+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release Notes
 All notable changes to this project will be documented in this file.
 
+
+## 0.24.0 - Feb 6, 2020
+### Featured
+- Auth support
+
+### New
+- NEW: --auth-endpoint (or AUTH_ENDPOINT env) config option. Specify address of auth service.
+- NEW: `authorization` optional header added to specify access token.
+- NEW: `auth` optional parameter added to all GraphQL queries to specify access token in GraphQL playground.
+
 ## 0.23.0 - Feb 5, 2020
 
 ### New